fix(tests): Moved Graph testing to correct dir
