+++ src/graph.ts
@@ -1,14 +1,12 @@
 import type { Subject, UserId } from "./acl.ts";
-import { Obj, Relation, UserSet } from "./acl.ts";
-import { describe, it } from "node:test";
-import { strict as assert } from "node:assert";
+import { Obj, Relation } from "./acl.ts";
 
 type Vertex = Obj | UserId;
 
 /**
  * Graph containing relations and vertices between them.
  */
-class Graph {
+export default class Graph {
     vertices: Vertex[];
     edges: Relation[];
 
@@ -56,46 +54,3 @@ class Graph {
         return users;
     }
 }
-
-describe("A graph", () => {
-    const mortenEhr = new Obj("EHR", "Morten's");
-
-    const læge = new Obj("Group", "Læge");
-    const overLæge = new Obj("Group", "Over Læge");
-
-    const Bob = 0;
-    const Alice = 1;
-    const Knud = 2;
-    const Gertrud = 3;
-    const Martin = 4;
-
-    const edges: Relation[] = [
-        new Relation(mortenEhr, "viewer", new UserSet(læge, "member")),
-        new Relation(læge, "member", new UserSet(overLæge, "admin")),
-        new Relation(læge, "member", new UserSet(overLæge, "member")),
-        new Relation(overLæge, "admin", Bob),
-        new Relation(overLæge, "member", Alice),
-        new Relation(læge, "member", Knud),
-        new Relation(mortenEhr, "viewer", Alice),
-        new Relation(mortenEhr, "viewer", Gertrud),
-        new Relation(overLæge, "ASS!!", Martin),
-    ];
-
-    it("should resolve all subjects", () => {
-        const graph = new Graph([], edges);
-
-        const users = graph.resolveSubjects(new UserSet(mortenEhr, "viewer"));
-
-        assert.deepEqual(users, new Set([Bob, Alice, Knud, Gertrud]));
-    });
-
-    it("should handle loops", () => {
-        const with_loop = edges.concat([
-            new Relation(overLæge, "member", new UserSet(læge, "member")),
-        ]);
-        const graph = new Graph([], with_loop);
-
-        const users = graph.resolveSubjects(new UserSet(mortenEhr, "viewer"));
-        assert.deepEqual(users, new Set([Bob, Alice, Knud, Gertrud]));
-    });
-});
